Suppress test timeout on HPE Apollo
The test empty-chpl-remote-taskspawn times-out on the HPE Apollo because of
known performance problems with remote task creation on that platform
(https://github.com/Cray/chapel-private/issues/5385). Suppress the failure of
the test on that platform until the underlying problem is fixed. Reduce the
timeout so the failures don't take so long, the default timeout is much too
long even on platforms on which the test succeeds.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>