Move the versionButton.php code to the www repo from the code repo
In discussions with Ahmad about how to streamline the release
procedures, we concluded that the changes to versionButton.php are
more closely related to events on the website (creating a new docs
snapshot; updating the links to make that snapshot the default) than
to the code directory.  For that reason, replacing the file that was
here to a symbolic link in the parent directory (stored in chapel-www)
makes sense, and we'll also do the reverse on the web repo.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>