Attempt to address valgrind timeout on spin version of this test
The valgrind executions of this test have been timing out with some
frequency lately after some system upgrades in the spin version of the
algorithm.  I think that, to be correct, the spin versions should be
yielding the processor, so here I'm updating the test to do just that
in hopes that it helps.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>