Fix a const-ness issue in cHelper.c
Until yesterday, interopWithC.chpl had not been getting tested with
the C back-end because its skipif was skipping if LLVM wasn't being
used as the back-end compiler (rather than if LLVM was configured
so that 'extern' blocks could be used).

After fixing this in https://github.com/chapel-lang/chapel/pull/26783
yesterday, we started getting a warning-as-error for certain back-end
C configurations (e.g., gcc, but seemingly not clang) due to
'const'ness being discarded when passing an array out to the external
'sumArray' call.  That call wasn't modifying its array argument, so in
this PR, I've added a 'const' qualifier to the routine and its
prototype, which makes the issue go away.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>