Update UDP docs to mention "DNSLookup" error and `CHPL_RT_MASTERIP`
This error was sufficiently unfamiliar to me that it would not have
occurred to me to apply this fix, and I didn't see any indication
of it in our docs, so added this section to capture it and the fix.

I've had this edit sitting in my tree for a few days now, but am
pushing it tonight to fire off another smoke test.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>