Don't free getenv result
getEnvInt calls createAdoptingBuffer with the C string returned by getenv. It
then returns the string cast as an int which frees the string. getenv returns
a string constant, so this results in the free of an address that was not
allocated, resulting in memory corruption. Calling createCopyingBuffer does
not free the string when getEnvInt returns.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>