Mark the new method as unstable
There was some uncertainty in the team meeting today about whether `int` was
the right return type for the method.  There was also a suggestion to use a type
argument for users to adjust it if they want.  I'm planning on opening an issue
about the latter suggestion (as well as if we should support a
`totalNanoseconds` method), but in the meanwhile we should mark the method as
unstable.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>