Turn off Apollo playground for now; set up for Qthreads as next config
This turns off the performance playground on Apollo, which had been
testing the new GASNet, hadn't shown any major surprises, and that
GASNet is now on main.

This'll be the first run with it off since Ahmad's
https://github.hpe.com/hpe/hpc-chapel-ci-config/pull/1369 was merged,
so it'll be interesting to see whether it stays blue as expected.

Meanwhile, I've also updated the playground to test Qthreads on
multilocale, and will plan to enable that after a night or two of
having it off to start gathering data there with an eye towards 2.5.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>