Improve portability of leak testing
I forgot to take into account that the amount leaked could reflect:

* CHPL_COMM=none vs. !=none
* 32- vs 64-bit linux
* C compiler choices

which made my leaking.bad file a bit too specific to where I tested
it.  This expands the prediff to not just filter the addresses being
leaked but any other integer preceded by 7 spaces, which happens to
cover the number of bytes per domain, and total, leaked, but also the
count of 1.  This all seems acceptable, though.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>