Add a new docs-release Makefile target
With the 'versionButton.php' reorganization that took place in PR
care with the versionButton.php file.  This PR adds a new target to
the developer and doc/ Makefile called 'docs-release' that sets the
versionButton.php file up to be the symbolic link we need it to be in
the release (which differs from the one we need it to be when we're
pushing the 'main' version of the docs to 'docs/main' on the live
website).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>