Revert get_string(); specialize conversion to field in postFold
Unfortunately, my previous attempt to modify get_string() failed to
realize how broadly it's used, and as a result, it broke a few other
tests ('extern/bradc/require/requireNonString.chpl' and
'library/packages/UnitTest/PrimitivesError/GetByName_WrongArgs.chpl').
Here, I've reverted that change and specialzed the logic that converts
arg #2 in a PRIM_GET_MEMBER[_VALUE] call to handle both string
literals and literal VarSymbol*'s that refer directly to fields.

Is this too specialized / distinct from this PR?  Should I just give
up and use a C string literal instead?  Perhaps, but it feels really
regrettable to me that we're still relying on that pattern in the IR,
so this is my baby step to try and not contribute to continuing that
practice.  With more effort, we could probably/possibly convert more
creations of these expressions to use the field directly, but that
would definitely feel like a step too far for this PR.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>