Expand testing to support only half-nan values and fix pre-existing bug
I realized I should add some tests of complex values who were just nan
in one component or the other and found that the string-to-complex
casts didn't work in the event that the imaginary portion was nan.
But now I think this was another instance of
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=48760 that could be fixed
using the same technique as in
https://github.com/chapel-lang/chapel/pull/24184, so I did that and
then added a test to lock in the behavior, which didn't work prior to
this commit.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>