Fix a bug in handling of array 'type' fields from #26690
This resolves a bug introduced in yesterday's #26690 fix to call
deinit() on fields when init/postinit throws.  Specifically, the net I
cast was a bit too wide, and we were trying to call deinit() on 'type'
fields that are arrays because they look like record fields (yet don't
support a 'deinit()' call because we don't suppot 'deinit()' on
types).

The fix here is simple, which is to just look for the presence of the
"type variable" flag on the symbol and to skip over such cases when
we're inserting the deinits.

This showed up in a pair of Python tests:

* test/library/packages/Python/correctness/arrays/nestedArrays.chpl
* test/library/packages/Python/correctness/arrays/numpyInterop.chpl

though they didn't get reported in nightly testing for some reason
(Jade is looking into why).  Here, I've added a specialized test to
lock in this behavior going forward.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>