Always use a '+' sign if the imaginary is nan (nani)
Prior to this, I was finding that the output would differ when
compiled on my Mac with clang vs. on linux with gcc, where the 'nani'
term would wiggle between '+' (on Mac) and '-' (on Linux).  My
understanding is that nans don't really have signs, to here, I'm
just always using '+' if the imaginary component is nan.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>