Remove the check that the compiler doesn't insert deinit's of records
PR #26990 added record deinit() calls as part of its work to
deinitialize classes in the event of throwing initializers / 'new'
calls.  This caused all such tests to fail when compiling with
--verify, as we have had a pre-existing check for some time to ensure
that the compiler-generated code does not delete/destroy/deinit
records.

In this PR, I'm removing the check to permit these calls to go
through, since the point of 'deinit()' is to support the compiler's
ability to call it, like the calls inserted in #26990.

Clearly, the compiler is calling deinit already when destroying things
in other code paths, but DavidL and I postulated that it's only doing
it through chpl__autoDestroy().  If correct, that suggests my PR
could've called chpl__autoDestroy() instead, but I view that call as
an unfortunate carryover from early days prior to completing
initializers and deinit(), and think it's clearer and more
straightforward to call deinit() directly when that suffices, like
here.

Note that user calls to deinit() are prohibited by other means, so
this doesn't change the user's ability to (not) call deinit.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>