Don't bother storing directly silenced errors in queries
Query errors need a handle on their errors for two reasons:

1. For re-emitting them when a generation changes,
   and a query result is re-used. In this case, we don't
   re-run the computation, but we do show the errors.
2. For emitting them when a query was previously
   executed as part of `runAndTrackErrors`, and is now
   being executed directly. In this case, errors
   were stored but not emitted. Now that the query
   is being "just" executed, they should be shown.

However, if the _current_ query is runnign 'runAndTrackErrors',
the errors should be shown in neither of the two contexts
above. If the result is re-used, we previously used
the error as part of the computation, and didn't emit it,
so we shouldn't emit it upon re-use. If the query
was inside another 'runAndTrackErrors', well, now we have
two reasons not to emit the error.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>