Add missing `const` qualifiers from `char*` declarations
The code here is longstanding, but has recently started getting tested
in GASNet's test configurations as a result of #26236, which makes
greater use of slurm by default.  They compile in developer mode to
get maximal errors and warnings, and that tripped a warning for some
`char* ... = "literal";` declarations that should've been `const
char*`.

This suggests that we should potentially be dialing up warnings in our
own nightly testing, which is somewhat thematically related to
https://github.com/Cray/chapel-private/issues/3138

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>