Store when-statement case type normally, store bool in AssociatedAction
Previously we were storing the boolean result of a when-statement's '=='
operator in the ResolvedExpression for the case-expression. This was
inconsistent with how we do things elsewhere, and caused complications
in the typed converter.

This commit adds a QualifiedType to AssociatedActions and stores the
boolean result there. Storing this result is particularly important in
the case where it is param-true or param-false.

This commit also adds a 'getAction' method to ResolvedExpression to
facilitate fetching the 'COMPARE' action.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>