Update more files to handle 16-bit floats
The main thing I'm aware of that's completely busted currently is
that my casts to complex rely on .re and .im which themselves
currently rely on casts in order to use chpl_crealf() implemented
in terms of __builtin_crealf, as I haven't found a 16-bit equivalent.
This results in an infinite recursion in the compiler.  I think
the way to break it is to enable something like the commented-out
code in chpl_anycomplex.re/.im

I think there are also updates needed to the CPP_IS_LAME macros
in frontend/lib/immediates/num.h, but I don't quite follow the
second of these yet.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>