First stab at writing .reshape() as a method to do an aliasing reshape
This is an implementation of a .reshape() method that creates a new
array that provides a view on an existing array's data using a new
array record.  That is, the original array and its domain remain
unchanged, a new array is created with the new shape and indices, but
it aliases the original array's data rather than doing a deep copy of
it.  If the result is stored in a new 'var', then a deep copy is made
at that point, but if it is stored in a 'ref' or used directly, no
deep copy is made.

Currently, this is only supported for default rectangular domains and
arrays, and it may only make sense to provide it for non-distributed
arrays in the long-term since viewing data "in-place" in a distributed
setting is arguablya bit weird.

In addition to a basic size check, we also do a check to see whether
the dimensions generally align, to avoid a challenging problem to
debug that a user brought up online.

TODOs:

- [ ] Is it too weird to have a standalone reshape() that copies and a
      method reshape() that does not?  And/or should we consider making
      the standalone reshape( ) not copy?

- [ ] Add a config that permits turning off the warning?

- [ ] Add more tests of error conditions, specifically size mismatches

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>