Add support for unbounded ranges; try (but fail) to avoid always aliasing
On the second point, I thought (but apparently had never checked)
that

var B = A.reshape(...);

would cause B to not alias A, but apparently it does...  Presumably,
this is because the compiler detects that the array being returned
by the routine is dead, so its memory can be stolen?  Here, I tried
adding a pragma to the routine to stop that, but it doesn't seem to
have worked?

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>