Skip resolving candidates for 'out' intent in more cases
If a variable clearly has a type or initializer, or is
declared as part of a loop, there's no way that a call
can be used to initialize it via the `out` intent.
As a result, there's no reason to avoid skipping a call
if that variable is unknown: in those cases, it's unknown
due to an error or something similar.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>