Rename getStarTuple, fix getQualifiedTuple var-arg-ness
Renames ``TupleType::getStarTuple`` to ``getVarArgTuple``, to better
reflect the manner in which the method is used today.

While making this change, I noticed that ``getQualifiedTuple`` was
setting all tuples it made to be var-arg tuples, which was certainly
incorrect. To fix this, I added an optional boolean argument that is
manually set in the right places.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>