Fix documentation issue with exponentiation in the spec, pointed out by Damian
Damian noticed that we were accidentally using a `+` operator in an example
involving exponentiation, when we should have been using `*`.  He also pointed
out that the documentation mentioned a `pow` function but that he was having
trouble finding it - I think this is because the `**` operator used to be a
function named `pow` a long time ago and we must have missed updating this
portion of the spec when it became the `**` operator.

Rectify both of those errors, and use a wordsmithing change he suggested in the
previous paragraph.  I also validated that the two code snippets being compared
did give the same result now for at least one set of `a, b, y` values.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>