Ensure formal Parent(X=?) with actual Child(A, B) becomes Parent(A)
Previously, Dyno always instantiated parent class formals with the
child class, which meant that the `Parent(X=?)` became Child(A, B),
but this breaks disambiguation for generic formals. Specifically,
after instantiations, both of the following overloads would look
the same:

```Chapel
proc (Parent(?)).foo() {}
proc (Child(?)).foo() {}
```

In both cases, the `Child(A,B)` actual would be used as the
instnatiating type, which would mean the two signatures
being compared will be the same. This PR adjusts the logic
to find the common parent (if any), and use that for instantiation.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>