Avoid special-casing compiler for various array types.
The `dom` and `eltType` fields are [required as part of the DSI](https://chapel-lang.org/docs/technotes/dsi.html#id11).

In practice, they are relied on in various method on the `_array` builtin record:

https://github.com/chapel-lang/chapel/blob/47946ac16e576a8d92863f8cffb4efebccb2ff96/modules/internal/ChapelArray.chpl#L832-L834

https://github.com/chapel-lang/chapel/blob/47946ac16e576a8d92863f8cffb4efebccb2ff96/modules/internal/ChapelArray.chpl#L819-L820

So, instead of special casing various well-known array classes to extract
their `dom` and `eltType` fields, simply adjust the InitResolver code to
search for these required fields until they are found.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>