Add a test of an extern proc with a default that depends on a previous arg
While discussing the lack of productivity in calling out to external
procedures that take a pointer + size (say) and how that feels
old-fashioned for Chapel, it occurred to me that it might work to have
the size argument take a default value that was based on the previous
argument, like 'size = arr.size'.  This seems to work, as hoped, and
I couldn't find an existing test with the pattern, so adding this one
to lock it in and be aware of when/whether it stops working.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>