Disallow partial instantiations for 'this'
This is done by marking instantiations with partially-generic 'this'
as needing instantiation still (indicating they did not receive enough
info to instantiate). This marks them as "unknown" when computing the
return type, and is used by `new` handling to produce Unknown as well.

This way, if an instantiation error occurrs as part of InitResolver,
thew `new` expression is marked `Unknown` instead of becoming a generic
type that wreaks havoc on the rest of resolution

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>