Fix segfault when handling `in` or `var` loop intents
Variables as `with (in x)` are initialized from their outer counterpart.
This means that an initializer is invoked. However, the initializer
currently expects something like an assignment, and relies on there
being a right-hand-side AST. In the `in x` case, there is no such AST.
Adjust the code to gracefully handle that situation.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>