Fix generated methods for fields with same name as type
Specifically, cases like the following:

```Chapel
record myRec {
  var myRec: int;
}
```

Prior to this commit, the code generated:

```Chapel
record myRec {
  var myRec: int;

  // notional
  proc myRec.init(myRec: int) { /* ... */ }
}
```

This caused conflicts when resolving `init`. However, when a user
actually writes an initializer, they are effectively doing so without
mentioning the type:

```Chapel
record myRec {
  var myRec: int;

  // desugared a bit
  proc init(this, myRec: int) { /* ... */ }
}
```

The resolver is already set up to deal with a no-type no-init `this`,
because that's how user code is written (above). So, the type expression
introduced in the builders only serves to cause errors in the presence
of similarly-named fields. So, just remove it.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>