Handle in/inout formals when promotion occurs in call-init-deinit etc.
Prior to this commit, call-init-deinit would always resolve
`init=` or `=` using the exact type of the actual. This
is not the appropriate thing to do when promotion is involved,
since it's the yielded type / scalar type that's being copied.
This commit adjusts VarScopeVisitor's API to behave properly
in this case (including, but not limited to, disabling
copy elision for promoted calls)

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>