Add a flag for editions to the compiler, and only allow specific values
Current supported values are "2.0" and "pre-edition".  We're nowhere near close
making a new edition, but when we are, that function will definitely need to
change.

Since we anticipate the possibility of having to support parsing changes in
editions in the future, we must check the edition values when we set the flag
and no later.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>