Fix bad interaction between generics with defaults and init=
Some types in Chapel have `init=` copy a type field from the
RHS unless they have a different instantiation for the current field.
This looks something like this:

```Chapel
proc init=(rhs: T(?)) {
  this.field1 =
    if this.type.field1 == ? then rhs.type.field1 else this.type.field1;
  this.field2 =
    if this.type.field2 == ? then rhs.type.field2 else this.type.field2;
}
```

One such type is `list`, which has a `parSafe` field, which it copies
from the RHS unless it was explicitly specified (via type expression)
to have a different value.

This had a bad interaction with `computeReceiverTypeConsideringState()`
and our setup code. Prior to this commit, our setup code stripped
substitutions provided to `init=`, and initializes the state with defaults
from the type. This means that as soon as the first `compute...` is hit,
the receiver's type get initialized _with the type's default_,
rather than the one from the type expression. Code that relies on
the logic above, which considers `this.type.field1`, then incorrectly
initializes the field.

This commit changes the implementation to initialize the state with
substitutions from the type expression, and ignore the default fields.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>