Add dyno start at supporting edition attribute
Allows the attribute to be applied to symbols, and specifies "first" and "last"
as potential arguments associated with it, as well as other rules about what is
not allowed.

Doesn't do anything with the attribute just yet, and also doesn't allow it to be
supported in the production compiler.  Those are next steps

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>