Reduce special cases for types that do not need init/deinit
Right now we do not sufficiently account for all cases in which a type
does not need init/deinit, meaning that sometimes we skip checks and
others we don't. Instead of incomplete skipping of checks and resolution
for the sake of simplicity, let's just make it actually work for such
types.

This also helps simplify things for typed conversion, as this commit
will cause assignment operators to be resolved and stored.

Requires a change from using 'idToParentId' to 'ID::contains' to account
for usage in nested calls.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>