Fix FindElidedCopies::handleDisjunction when param-true conditional occurs
Previously we had some special case logic for frames when it was
statically determined that the frame would be taken. This was faulty
logic firstly because it failed to account for all previous frames,
which might be executed first (e.g. in the case of a 'select').
Secondly, the implementation tried and failed to recreate the logic
below, which had gotten somewhat out of sync over time.

This can be fixed by updating VarScopeVisitor's invocation of
'handleDisjunction' to correctly pass 'total', which indicates whether
any branch will always be taken. In doing so, we can get rid of the
special-case logic.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>