Move the edition check earlier in resolution
It was failing to trigger where it was if the edition version was the only one
available (because we don't need to disambiguate if there's only one candidate).
So move the check to visible functions.

Unfortunately this currently causes the error message to trigger twice.  Still
need to investigate that

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>