Dyno: continue resolution after 'throws'
Evidently this is necessary, since we have existing examples that
rely on return type inference in cases like `throw; return`.

To make this work, adjust the BranchSensitiveVisitor and its
related types to distinguish between "throws" and "returns",
and adjust the `isDoneExecuting` check to not return true
for throws. Then, for analyses that DO ignore code after
throwing etc., broaden the early-return check to include
throwing in addition to `isDoneExecuting`.

This has the net effect of making the Dyno resolver proceed
past `throws` statements, resolving expressions etc. in
`return` and `yield`s, as well as making return type inference
do something similar.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>