Make record docs of init= testable
As pointed out in #27249, we have not been testing a few of our spec
docs for records, and as a result, they silently broke when we added
the requirement that a mixed-type init= also define a cast between
those types.  This adds the missing casts and makes the remaining
examples testable.

Not done in this commit, I haven't added the rationale for why the
cast is required, because I haven't quite remembered it yet, so am
needing an assist or more time.

TODO:
- [ ] address TODO by adding a rationale for needing the cast

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>