Make 'make docs' clobber chpl-venv, which is often required in practice
This is in lieu of actually getting the 'make docs' dependencies
correct for now, where Jade outlines some of the challenges in the
following quote/paraphrase, and the Makefile comment referred to:

> `make docs` can break if, at some point chapel-py was built and then
libChplFrontendShared was removed, breaking the chapel-py
install. There are definitely some weird errors than can occur when
you try to selectively clobber and chapel-py is involved, because the
dependencies aren't quite right due to technical challenges. See
third-party/chpl-venv/Makefile:96. I would love to see this fixed, but
it's a big pain and the minor dev annoyance has not been worth the
time investment

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>