Add a mechanism for auto-numbering and selecting overloads.
The OverloadSelector is a template class that always maintains a fixed
order when multiple overloads are present, and allows one to map the
index of an overload to its generator.

The `invoke` method, given a particular overload's builder, invokes that
builder with an argument corresponding to that builders ordinal. This
ordinal is always the same; this way, each builder logically corresponds
to a particular overload. The intention is for the builder to
incorporate the ordinal into the IDs of the ASTs it generates. This can
be used with the `select` method.

The `select` method takes a runtime ordinal value and calls the builder
with the corresponding ordinal. This can be used when mapping IDs
produced by a builder back to their builder to retrieve the AST.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>