Add a test that "default" is not valid in the attribute
Since it's an option for the compilation flag, someone not thinking carefully
might try to use it when locking a change to a particular edition.  It might be
worth specializing that message for this particular value, but I'm not doing
that right now.

Anyways, this shouldn't be allowed because the default edition is expected to
change over time.  That means that if we change the default while still
maintaining 2.0, suddenly we could encounter resolution conflicts that wouldn't
be obvious unless we were explicitly testing that particular call.  No bueno.
An edition's behavior should be the same throughout its life cycle, *including*
when it becomes the default edition.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>