Changed binding to architectural features
Changed the logic in the previous commit so that co-locales are bound to an
architectural feature if the number of non-empty instances of that feature
equals the number of co-locales, and if the non-empty instances have an equal
number of cores. For example, the Apple M1Pro has three L2 caches, one of
which has efficiency cores and other two have performance cores. With this
change "-nl 1x2" will bind the two co-locales to the two L2 caches with
performance cores, and ignore the L2 cache that has efficiency cores.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>