First attempt to use editions + standalone procedure
Mostly seems to be working except that versions that take ranges
generate an error about being unable to return a scoped variable.
My guess is that this is more due to the use of procedures rather
than methods rather than being at all due to editions.  And/or
maybe `pragma "reference to const when const this"` was keeping
things working in the method case?

TODO:
- [ ] Does removing the pragma from the previous version cause the
      methods to fail in the same way?
- [ ] Does duplicating the code logic into the routine help?
      (though doing it for each branch of the conditional is not good...

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>