Add 'BaseAST::canMutateEarly()' and begin adjusting problematic passes
The predicate 'wasResolvedEarly()' is not enough to determine if we
can mutate some AST or not - things contained in e.g., a module are
allowed to be mutated because (at least for now) modules can contain
a mix of early and late resolved symbols.

Adjust some problematic passes that were causing the crash to trigger.
For some, just skip mutation. In at least one case adjust the code
the typed converter is outputting. Right now the branch is stuck on
'ReturnByRef', which is applying to early-resolved code. I think the
solution here will be to move that transform to the other side of the
"dividing line" by considering it to occur "after" 'callDestructors'.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>