Fix a bug in returning reshape() expressions
One of the follow-ups I had lined up for the new aliasing reshape()
capability was to determine why one of the four cases in
returnReshape.chpl wasn't working even though the others were.
Specifically, this call:

```chapel
proc bar() {
  return reshape(A, {1..2, 1..2});
}
```

was getting a lifetime checking error about returning a scoped
variable, when other similar calls like:

```chapel
proc foo() {
  return reshape(A, 1..2, 1..2);
}
```

were not.

Digging into this a bit more, I'm still not entirely certain why the
two cases were resulting in different behavior, though I did find that
the foo() routine should have been getting a similar error, in that
capturing its result into a variable and modifying it:

```chapel
var B = foo();
B[1,1] = 5;
```

was resulting in a modification to 'A' as well, even though returning
an array and storing it in a variable should result in a deep copy.

Once I realized this, it didn't take long to determine that this is
another place in the compiler where the current logic determines the
existence of an array view by virtue of the expression's type.
Updating it to look for the new FLAG_IS_ARRAY_VIEW flag fixed the
aliasing issue while also resolving the original error about returning
a scoped variable.

While here, I updated the test to make it check that none of the
modifications to the returned arrays resulted in a modification to the
original.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>