Add cast from real to bigint
Implemented using the pre-existing .set() method.

I think there's a little bit of a question around whether we'd want to
support init= from a real to a bigint as well.  An argument against
would be that we don't normally support init= from reals to integers.
An artgument "for" might be that bigints aren't typical ints...  But I
wasn't sure I believed that argument enough to propose doing it as
part of this PR.

Resolves #27390

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>