Close memory leak with PrivateDist
The new tests I added allocated new instances of 'privateDist' which
we don't technically prevent, and the docs suggest is reasonable.
However, this caused leaks because PrivateDist is labeled as a
singleton which prevents the implementation class from being freed.

Given that multiple instances can be created, I don't think it makes
sense to consider it a singleton as the code traditionally has.  It
may be that we used to have all PrivateDists share the same
implementation, and that I broke or changed that when I converted it
to a record, but I'm not certain.

This singleton marker was added in
https://github.com/chapel-lang/chapel/pull/4762 and
test/parallel/taskPar/sungeun/private.chpl was used as the rationale,
yet it seems to be passing, so I'm not sure that original motivation
holds any longer.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>