Update 'Select' conversion to work with branch sensitive traversal
Calling 'branchSensitivelyTraverse' for ``Select``, and relying on
traversal seems to make things notably more complicated than this
top-down whole-node approach. When generating a ``Select`` for
production AST we only want to generate the select's expression call
once, so we'd need to store that somewhere. It's also trickier to manage
cascading CondStmts from a bottom-up view that we'd have in a
pure-traversal setting.

Instead, we update the current implementation to correctly enter and
exit scopes as needed. Otherwise, traversal would use the overall
``Select``'s frame as a reference when deciding whether a subtree was
'dead', which isn't correct when considering individual ``When``
statements. For example, if all ``When`` statements returned, the
``Select``'s frame would indicate ``true`` when ``isDoneExecuting`` was
invoked, therefore skipping the desired node.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>